update examples section
