+++ examples/01-twin_examples/README.txt
@@ -1,0 +1,5 @@
+Twin evaluation examples
+=========================
+These examples show how to use the evaluation workflow in a TWIN file.
+Included are multiple workflow demonstrations such as single evaluation of a dynamic model, parametric sweep analysis
+(static/dynamic models), evaluation with states saving and loading.
