+++ examples/02-tbrom_examples/README.txt
@@ -1,0 +1,5 @@
+TBROM evaluation examples
+=========================
+These examples show how to use PyTwin to evaluate field results from TBROM included in a TWIN file.
+Multiple workflows are demonstrated such as image and field data generation (with possible processing of data to other
+Ansys tools like Fluent or Mechanical), TBROM evaluation with input field parameterization, output field visualization for detailed post processing.
