+++ examples/README.txt
@@ -5,9 +5,3 @@ Examples
 =========
 End-to-end examples show how you can use PyTwin. If the ``pytwin`` package is installed on your machine,
 you can download these examples as Python files or Jupyter notebooks and run them locally.
-
-.. toctree::
-   :maxdepth: 1
-   :hidden:
-
-.. evaluate/README.txt
