--- examples/evaluate/README.txt
@@ -1,12 +1,0 @@
-Twin evaluation examples
-=========================
-These examples show how to use the evaluation workflow in a TWIN file.
-Included are multiple workflow demonstrations such as these:
-
-- Single evaluation of a dynamic model
-- Parametric sweep analysis of a dynamic model
-- Parametric response surface ROM evaluation
-- Scalar dynamic ROM simulation with state saving and loading
-- Field ROM evaluation and results visualization from the twin runtime and postprocessed in Ansys Fluent
-- Field ROM evaluation and the transfer of results as inputs stress analysis in Ansys Mechanical
-- Field ROM evaluation with inputs parameterized by both scalar and field data, and snapshot generation on demand
